+++ app/client/src/components/editorComponents/ActionNameEditor.tsx
@@ -2,7 +2,6 @@ import React, { memo } from "react";
 import { useSelector } from "react-redux";
 
 import { useParams } from "react-router-dom";
-import styled from "styled-components";
 import EditableText, {
   EditInteractionKind,
 } from "components/editorComponents/EditableText";
@@ -11,9 +10,12 @@ import type { AppState } from "@appsmith/reducers";
 
 import { saveActionName } from "actions/pluginActionActions";
 import { Spinner } from "design-system";
-import { Classes } from "@blueprintjs/core";
 import { getAction, getPlugin } from "@appsmith/selectors/entitiesSelector";
-import NameEditorComponent from "components/utils/NameEditorComponent";
+import NameEditorComponent, {
+  IconBox,
+  IconWrapper,
+  NameWrapper,
+} from "components/utils/NameEditorComponent";
 import {
   ACTION_ID_NOT_FOUND_IN_URL,
   ACTION_NAME_PLACEHOLDER,
@@ -22,43 +24,6 @@ import {
 import { getAssetUrl } from "@appsmith/utils/airgapHelpers";
 import { getSavingStatusForActionName } from "selectors/actionSelectors";
 import type { ReduxAction } from "@appsmith/constants/ReduxActionConstants";
-
-const ApiNameWrapper = styled.div<{ page?: string }>`
-  min-width: 50%;
-  margin-right: 10px;
-  display: flex;
-  justify-content: flex-start;
-  align-content: center;
-  & > div {
-    max-width: 100%;
-    flex: 0 1 auto;
-    font-size: ${(props) => props.theme.fontSizes[5]}px;
-    font-weight: ${(props) => props.theme.fontWeights[2]};
-  }
-
-  ${(props) =>
-    props.page === "API_PANE"
-      ? `  &&& .${Classes.EDITABLE_TEXT_CONTENT}, &&& .${Classes.EDITABLE_TEXT_INPUT} {
-    font-size: ${props.theme.typography.h3.fontSize}px;
-    letter-spacing: ${props.theme.typography.h3.letterSpacing}px;
-    font-weight: ${props.theme.typography.h3.fontWeight};
-  }`
-      : null}
-`;
-
-const ApiIconWrapper = styled.img`
-  width: 34px;
-  height: auto;
-`;
-const ApiIconBox = styled.div`
-  height: 34px;
-  width: 34px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: 8px;
-  flex-shrink: 0;
-`;
 
 interface SaveActionNameParams {
   id: string;
@@ -71,7 +36,7 @@ interface ActionNameEditorProps {
     Right now, it's optional so that it doesn't impact any other pages other than API Pane.
     In future, when default component will be ads editable-text, then we can remove this prop.
   */
-  page?: string;
+  enableFontStyling?: boolean;
   disabled?: boolean;
   saveActionName?: (
     params: SaveActionNameParams,
@@ -121,7 +86,7 @@ function ActionNameEditor(props: ActionNameEditorProps) {
         isNew: boolean;
         saveStatus: { isSaving: boolean; error: boolean };
       }) => (
-        <ApiNameWrapper page={props.page}>
+        <NameWrapper enableFontStyling={props.enableFontStyling}>
           <div
             style={{
               display: "flex",
@@ -129,12 +94,12 @@ function ActionNameEditor(props: ActionNameEditorProps) {
             }}
           >
             {currentPlugin && (
-              <ApiIconBox>
-                <ApiIconWrapper
+              <IconBox>
+                <IconWrapper
                   alt={currentPlugin.name}
                   src={getAssetUrl(currentPlugin?.iconLocation)}
                 />
-              </ApiIconBox>
+              </IconBox>
             )}
             <EditableText
               className="t--action-name-edit-field"
@@ -154,7 +119,7 @@ function ActionNameEditor(props: ActionNameEditorProps) {
             />
             {saveStatus.isSaving && <Spinner size="md" />}
           </div>
-        </ApiNameWrapper>
+        </NameWrapper>
       )}
     </NameEditorComponent>
   );
