+++ app/client/src/components/utils/NameEditorComponent.tsx
@@ -11,6 +11,46 @@ import {
   ACTION_NAME_CONFLICT_ERROR,
   createMessage,
 } from "@appsmith/constants/messages";
+import styled from "styled-components";
+import { Classes } from "@blueprintjs/core";
+
+export const NameWrapper = styled.div<{ enableFontStyling?: boolean }>`
+  min-width: 50%;
+  margin-right: 10px;
+  display: flex;
+  justify-content: flex-start;
+  align-content: center;
+  & > div {
+    max-width: 100%;
+    flex: 0 1 auto;
+    font-size: ${(props) => props.theme.fontSizes[5]}px;
+    font-weight: ${(props) => props.theme.fontWeights[2]};
+  }
+
+  ${(props) =>
+    props.enableFontStyling
+      ? `  &&& .${Classes.EDITABLE_TEXT_CONTENT}, &&& .${Classes.EDITABLE_TEXT_INPUT} {
+  font-size: ${props.theme.typography.h3.fontSize}px;
+  letter-spacing: ${props.theme.typography.h3.letterSpacing}px;
+  font-weight: ${props.theme.typography.h3.fontWeight};
+}`
+      : null}
+`;
+
+export const IconWrapper = styled.img`
+  width: 34px;
+  height: auto;
+`;
+
+export const IconBox = styled.div`
+  height: 34px;
+  width: 34px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-right: 8px;
+  flex-shrink: 0;
+`;
 
 interface NameEditorProps {
   checkForGuidedTour?: boolean;
