+++ app/client/src/ee/pages/Editor/IDE/MainPane/useRoutes.ts
@@ -1,0 +1,3 @@
+export * from "ce/pages/Editor/IDE/MainPane/useRoutes";
+import { default as useCE_Routes } from "ce/pages/Editor/IDE/MainPane/useRoutes";
+export default useCE_Routes;
